Performs a statistical analysis of 2x2 bioequivalence study

beNCA returns text output of a statistical analysis of 2x2 bioequivalence study. Analysis of variance, least square means, confidence interval, and sample size will be calculated for AUClast and Cmax. Wilcoxon Signed-Rank Test and Hodges-Lehmann Estimate will be performed for Tmax.

beNCA(concData, SUBJ = "SUBJ", GRP = "GRP", PRD = "PRD", TRT = "TRT",
  method = "kbe", ...)

Arguments

SUBJ
Subject ID, any data type

GRP
column name in which information of "RT" or "TR" exists.

PRD
column name in which information of 1 or 2 exists.

TRT
column name in which information of "R" or "T" exists.

method
kbe by authors or nlme package uploaded on CRAN

Value

returns text results of statistical analysis of 2x2 bioequivalence study including and beNCAdataset.csv file in the working directory which can be run in SAS.

Details

Reference: Shein-Chung Chow, Jen-pei Liu. Design and Analysis of Bioavailability and Bioequivalence Studies, 3rd ed. 2008. (ISBN:9781584886686)

Examples

file <- system.file('example', 'beConc.csv', package = 'ncarbe')
concData <- read.csv(file, as.is = TRUE)
beNCA(concData)

#> 
#> 
#> [AUClast]
#> 
#> $`Analysis of Variance`
#>                        SS DF        MS        F           p
#> SUBJECT        35.7817072 35 1.0223345 1.854676 0.037320327
#> GROUP           1.1390308  1 1.1390308 1.117900 0.297825788
#> SUBJECT(GROUP) 34.6426764 34 1.0189022 1.848449 0.038831903
#> PERIOD          4.1989873  1 4.1989873 7.617625 0.009244748
#> DRUG            0.7021705  1 0.7021705 1.273848 0.266947114
#> ERROR          18.7414811 34 0.5512200       NA          NA
#> TOTAL          58.9756619 71        NA       NA          NA
#> 
#> $`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                     0.2338411        0.55122
#> Coefficient of Variation, CV(%)      51.3267693       85.75365
#> 
#> $`Least Square Means`
#>                 Reference Drug Test Drug
#> Geometric Means       31.20441   25.5401
#> 
#> $`90% Confidence Interval`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio   0.6062815      0.8184771     1.10494
#> 
#> $`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size           96                      6538
#> 
#> 
#> 
#> [Cmax]
#> 

#> Warning: NaN이 생성되었습니다

#> $`Analysis of Variance`
#>                        SS DF        MS         F          p
#> SUBJECT        22.1199814 35 0.6319995 0.6475704 0.89703843
#> GROUP           0.2047080  1 0.2047080 0.3175901 0.57675784
#> SUBJECT(GROUP) 21.9152734 34 0.6445669 0.6604475 0.88422171
#> PERIOD          3.6728929  1 3.6728929 3.7633843 0.06071806
#> DRUG            0.1136911  1 0.1136911 0.1164922 0.73497096
#> ERROR          33.1824629 34 0.9759548        NA         NA
#> TOTAL          58.9756619 71        NA        NA         NA
#> 
#> $`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                     -0.165694      0.9759548
#> Coefficient of Variation, CV(%)             NaN    128.5962569
#> 
#> $`Least Square Means`
#>                 Reference Drug Test Drug
#> Geometric Means       26.30688  28.51506
#> 
#> $`90% Confidence Interval`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio   0.7270811       1.083939    1.615946
#> 
#> $`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size          169                       298
#> 
#> 
#> 
#> [Tmax]
#> 
#> $`Wilcoxon Signed-Rank Test`
#>   p-value 
#> 0.3286283 
#> 
#> $`Hodges-Lehmann Estimate`
#>                            Lower Limit Point Estimate Upper Limit
#> 90% Confidence Interval       -2.00000         1.0000      4.0000
#> 90% Confidence Interval(%)    87.11261       106.4437    125.7748
#>