Performs a statistical analysis of 2x2 bioequivalence study

beNCA returns text output of a statistical analysis of 2x2 bioequivalence study. Analysis of variance, least square means, confidence interval, and sample size will be calculated for AUClast and Cmax. Wilcoxon Signed-Rank Test and Hodges-Lehmann Estimate will be performed for Tmax.

beNCA(concData, SUBJ = "SUBJ", GRP = "GRP", PRD = "PRD", TRT = "TRT",
  method = "kbe", ...)

Arguments

SUBJ
Subject ID, any data type

GRP
column name in which information of "RT" or "TR" exists.

PRD
column name in which information of 1 or 2 exists.

TRT
column name in which information of "R" or "T" exists.

method
kbe by authors or nlme package uploaded on CRAN

Value

returns text results of statistical analysis of 2x2 bioequivalence study including and beNCAdataset.csv file in the working directory which can be run in SAS.

Details

Reference: Shein-Chung Chow, Jen-pei Liu. Design and Analysis of Bioavailability and Bioequivalence Studies, 3rd ed. 2008. (ISBN:9781584886686)

Examples

file <- system.file('example', 'beConc.csv', package = 'ncarbe')
concData <- read.csv(file, as.is = TRUE)
beNCA(concData)

#> 
#> 
#> [AUClast]
#> 
#> $`Analysis of Variance`
#>                       SS DF         MS        F          p
#> SUBJECT        2.8102897 35 0.08029399 2.049048 0.01947896
#> GROUP          0.2811516  1 0.28115157 3.779609 0.06019307
#> SUBJECT(GROUP) 2.5291381 34 0.07438642 1.898290 0.03291404
#> PERIOD         0.2887407  1 0.28874073 7.368465 0.01035002
#> DRUG           0.1186721  1 0.11867210 3.028430 0.09086256
#> ERROR          1.3323243 34 0.03918601       NA         NA
#> TOTAL          4.5500268 71         NA       NA         NA
#> 
#> $`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                     0.0176002     0.03918601
#> Coefficient of Variation, CV(%)      13.3251642    19.99097469
#> 
#> $`Least Square Means`
#>                 Reference Drug Test Drug
#> Geometric Means       5047.026  4648.063
#> 
#> $`90% Confidence Interval`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio   0.8501323       0.920951   0.9976692
#> 
#> $`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size            8                        13
#> 
#> 
#> 
#> [Cmax]
#> 

#> Warning: NaN이 생성되었습니다

#> $`Analysis of Variance`
#>                        SS DF         MS         F         p
#> SUBJECT        2.85581816 35 0.08159480 0.9764709 0.5283290
#> GROUP          0.08840271  1 0.08840271 1.0861008 0.3046908
#> SUBJECT(GROUP) 2.76741545 34 0.08139457 0.9740747 0.5302983
#> PERIOD         0.04931289  1 0.04931289 0.5901430 0.4476669
#> DRUG           0.10908566  1 0.10908566 1.3054627 0.2612001
#> ERROR          2.84107119 34 0.08356092        NA        NA
#> TOTAL          5.85528790 71         NA        NA        NA
#> 
#> $`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                  -0.001083173     0.08356092
#> Coefficient of Variation, CV(%)             NaN    29.52142270
#> 
#> $`Least Square Means`
#>                 Reference Drug Test Drug
#> Geometric Means       791.1619     731.1
#> 
#> $`90% Confidence Interval`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio   0.8221796       0.924084    1.038619
#> 
#> $`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size           16                        26
#> 
#> 
#> 
#> [Tmax]
#> 
#> $`Wilcoxon Signed-Rank Test`
#>   p-value 
#> 0.3059991 
#> 
#> $`Hodges-Lehmann Estimate`
#>                            Lower Limit Point Estimate Upper Limit
#> 90% Confidence Interval       -0.10000         0.0300       0.405
#> 90% Confidence Interval(%)    92.18517       102.3444     131.650
#>